+++ plugins/clear.go
@@ -2,7 +2,6 @@ package plugins
 
 import (
 	"fmt"
-	"log"
 	"os"
 	"regexp"
 	"strings"
@@ -57,7 +56,6 @@ func (c *clear) clearMSGs(toDeleteID string, msg *dg.MessageCreate) error {
 		}
 	}
 
-	log.Println(msgIDs)
 	err = c.dgs.ChannelMessagesBulkDelete(msg.ChannelID, msgIDs)
 	if err != nil {
 		return fmt.Errorf("ChannelMessageBulkDelete failed: %s", err)