+++ plugins/plugins.go
@@ -11,6 +11,9 @@ import (
 const CommandDelay = 90
 const botDelay = CommandDelay * time.Second
 
+// Environment variables for the plugins package.
+const botOwner = "BOT_OWNER"
+
 // AkSession allows for tests to mock the discordgo session.Channel() method call
 type AkSession interface {
 	Channel(channelID string) (st *dg.Channel, err error)
@@ -29,6 +32,7 @@ func NewRecorder() *Record {
 	return &Record{LastReq: userMap, MinWaitTime: (botDelay)}
 }
 
+// CheckLastAsk checks the last time the user executed the specific command
 func (r *Record) CheckLastAsk(msg *dg.MessageCreate) (string, bool) {
 	last, found := r.LastReq[msg.Author.ID]
 	if found && time.Since(last) < (r.MinWaitTime) {